requirements.txt
langchain==0.0.250
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.25.0
streamlit-extras
streamlit-chat==0.1.1
tiktoken==0.4.0
pinecone-client==2.2.2
openai==0.27.10
